Add a `get-random-u64` function.
Rename `getrandom` to `get-random-bytes` and add a `get-random-u64`
function. It returns the same kind of random data, just in a different
format. A `u64` is more convenient for some use cases, and in this time
when returning a `list<u8>` may incur an allocation, returning a `u64`
avoids that overhead.

Thanks to @npmccallum for suggesting this!