Add an `insecure` interface.
Add an `insecure` interface, and split the seed function into its own
interface, producing three interfaces:
 - `random`, for secure pseudo-random numbers
 - `insecure`, for pseudo-random numbers that are not secure
 - `insecure-seed`, for supporting hash-map DoS protection

We've gone back and forth on whether such an API is desirable, but it's
now motivated by the need to implement things like JavaScript's
`Math.random`, which want to be random, don't want to run their own PRNG,
and don't need to be secure. And in fact, they specifically don't want
to be secure, because they want to allow hosts to make them
deterministic if they wish to.