ci: lint and publish 0.3.0 timestamp based rc's
The publish workflow for 0.3, unlike publish for main, is manually
triggered (vs being triggered by a release). It's easier to turnkey the
timestamp tags this way and the assumption is that we wouldn't want to
trigger "releases" for these while the churn rate is high. This also
aims to avoid creating PR's across repos that would require
manual approval.

Before a tag has been created the version value will be:
```bash
echo $VERSION
v0.3.0-rc-2025-07-22
```

Multiple revs in a single day will produce: v0.3.0-rc-2025-07-22-n